feat(tui): bank-aware memory panel — :bank selector for the 65816
The memory panel can view and edit 65816 banks 1-255: Source.ReadMemory widens
to a 24-bit address, refreshMemWindow anchors the window at MemViewBank:MemViewAddr,
:bank N selects the bank, and the panel header + row addresses show $BB:XXXX.
Bank>0 edits write through the Banked24 (bank 0 stays on the watchpoint chain).
MemViewBank persists as an additive v1.x state field. WithBanked24 shares the bus
with the in-process DAP server. Cross-bank disassembly (PBR-relative) is a
documented follow-up; disasm stays PC/bank-0 anchored.

Refs #505

+++ internal/tui/source_remote.go
@@ -243,10 +243,12 @@ func (s *RemoteSource) Flags() (FlagsSnapshot, error) {
 // (issue #451). The mirror is reconciled by RefreshMemory on every stop and
 // updated by #440 dirtyRanges during a run, so the panel needs no per-frame
 // wire round-trip — reading the mirror IS reading DAP-sourced bytes.
-func (s *RemoteSource) ReadMemory(addr uint16, count int) ([]byte, error) {
+func (s *RemoteSource) ReadMemory(addr uint32, count int) ([]byte, error) {
+	// Remote debuggees are 6502/NES (no 65816 host), so the mirror is bank 0;
+	// a 24-bit address masks into it.
 	buf := make([]byte, count)
 	for i := 0; i < count; i++ {
-		buf[i] = s.ram.Read(addr + uint16(i))
+		buf[i] = s.ram.Read(uint16(addr) + uint16(i))
 	}
 	return buf, nil
 }
